Improve compiler behavior around stack overflow caused by long binary expressions.
Fixes #5395.

Adjusted Binder, Optimizer, Flow Analysis and code generator to not use recursion to handle binary expressions nested on the left side of another binary expression.
Adjusted CodeGenerator to detect that stack overflow is about to happen while trying to emit IL for an expression and report a diagnostic pointing to the expression at fault instead of crashing compiler.